Add in basic functions for big-endian support (utendian.c)
Add functions to convert to little-endian (AcpiUtConvertHostIntToLet())
and to convert to big-endian (AcpiUtConvertLEToHostInt()).  This is instead
of using macros to hide the conversion -- for example, ACPI_MOVE_32_TO_32()
does that, and the macro can be rewritten to hide the conversion to or
from big-endian; the point is, it _is_ hidden, and maybe all you wanted to
do was move bytes.

With these functions, any endian conversion is explicit and visible.  Since
these can be null on little-endian machines, they also impose no overhead
on the common use case, just the big-endian case.

Note that big-endian support is primarily for the user-space tools such
as iasl.  In particular, QEMU on s390x relies on iasl to generate ACPI
tables for x86_64 virtual machines, and hence the need to big-endian
support.

Signed-off-by: Al Stone <ahs3@ahs3.net>