Re-enable big-endian support
Several code changes caused namespace initialization to occur a little
bit differently.  Make sure that (a) little-endian path names are used
and (b) that little-endian path names are being searched for; this
allows us to initialize the pre-defined names and then search for the
name \_GPE to make sure we've set things up correctly.

Signed-off-by: Al Stone <ahs3#redhat.com>