Add in basic infrastructure for big-endian support
This adds in some basic functions -- AcpiUtReadUint32(), to read a
UINT32 value in little-endian form and return it in host-native format,
along with AcpiUtWriteUint() that writes out an integer in host-native
format as a little-endian value.

But, to do that, I'm adding the functions in a new file: utendian.c.  So,
the header files need fixing, and the makefiles need to be sure to compile
the new code.

However, this sets things up for the future, where endian-aware code
can be added as the need is uncovered.

Signed-off-by: Al Stone <ahs3@redhat.com>