Support DSDT/SSDT in a big-endian world.
NB: this is a very large diff.  The problem is that ResourceTemplates
are treated differently during compilation and disassembly so each of
the resource types had code that needed to be touched directly.

In general, however, just reading or writing individual AML opcodes
wasn't that much of a change, and most of it was on the codegen side.

Signed-off-by: Al Stone <ahs3@redhat.com>