Correct DBG2 dump
When OemData is present in a Device subtable, the original code was
reading data to put in OemData from the beginning of the DBG2 table,
not from the subtable that was being built.  So, when dumping the
table, read from the proper place (the subtable) and not the beginning
of the DBG2 table.

Signed-off-by: Al Stone <ahs3@ahs3.net>