Use original DataTableRegion pointer for accesses
Currently the pointer to the table is cast to ACPI_PHYSICAL_ADDRESS and
later cast back to a pointer to be dereferenced. Whether or not this is
supported is implementation-defined.

On CHERI, and thus Arm's experimental Morello prototype architecture,
pointers are represented as capabilities, which are unforgeable bounded
pointers, providing always-on fine-grained spatial memory safety. This
means that any pointer cast to a plain integer will lose all its
associated metadata, and when cast back to a pointer it will give a
null-derived pointer (one that has the same metadata as null but an
address equal to the integer) that will trap on any dereference. As a
result, this is an implementation where ACPI_PHYSICAL_ADDRESS cannot be
used as a hack to store real pointers.

Thus, add a new field to ACPI_OBJECT_REGION to store the pointer for
table regions, and propagate it to AcpiExDataTableSpaceHandler via the
region context, to use a more portable implementation that supports
CHERI.